Update Arm_Gen
Offset wrist control wasn't 'rotating' the forearm as expected... made it so that both the wrist and offset wrist can rotate it at the same time using addDoubleLinear.

Also made the forearm rotation a bit nicer. It used to be exponentially decreasing in rotation. Now it is linear, rotate/1, rotate/2, rotate/4, rotate/6...